projects: split api calls get and submit
